requirements.txt
django
pre-commit
djangorestframework
markdown
django-filter
